laravel spatie permissions implemented

+++ app/Http/Controllers/CandidatureController.php
@@ -6,10 +6,11 @@ use App\Candidat;
 use App\docFile;
 use App\Formation;
 use App\Candidature;
-use PDF;
+use Barryvdh\DomPDF\Facade as PDF;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class CandidatureController extends Controller
 {
@@ -20,7 +21,7 @@ class CandidatureController extends Controller
      */
     public function index()
     {
-
+        abort_if(Gate::denies('Candidature_access'), 403);
         $candidatures = DB::table('candidatures')
             ->join('candidats', 'candidat_id', '=', 'candidats.id')
             ->join('formations', 'formation_id', '=', 'formations.id')
@@ -70,6 +71,7 @@ class CandidatureController extends Controller
      */
     public function downloadPDF($id)
     {
+        abort_if(Gate::denies('Candidature_PDF_download'), 403);
         $candidat = Candidat::where('id', $id)->first();
         $pdf = PDF::setOptions(['isHtml5ParserEnabled' => true, 'isRemoteEnabled' => true]);
         set_time_limit(300);
@@ -86,6 +88,8 @@ class CandidatureController extends Controller
      */
     public function showPDF($id)
     {
+        
+        abort_if(Gate::denies('Candidature_PDF_view'), 403);
         $candidat = Candidat::where('id', $id)->first();
         return view('pre-inscription.attestation')->with('candidat', $candidat);
     }
@@ -99,6 +103,7 @@ class CandidatureController extends Controller
     public function edit( $id )
     {
 
+        abort_if(Gate::denies('Candidature_edit'), 403);
 
         $candidat = Candidat::where('id',Candidature::where('id',$id)->first()->candidat_id)->first();
 
@@ -108,6 +113,7 @@ class CandidatureController extends Controller
     public function editValidation(Candidature $candidature, $id)
     {
 
+        abort_if(Gate::denies('Candidature_edit'), 403);
         $candidature = Candidature::findOrFail($id);
 
         ($candidature->valide == 1) ? ($candidature->valide = 0) : ($candidature->valide = 1);
@@ -143,7 +149,7 @@ class CandidatureController extends Controller
      */
     public function update(Request $request, Candidature $candidature)
     {
-        dd(\App\Candidature::all());
+        abort_if(Gate::denies('Candidature_update'), 403);
     }
     /**
      * Remove the specified resource from storage.
@@ -153,6 +159,7 @@ class CandidatureController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(Gate::denies('Candidature_delete'), 403);
         $candidature = Candidature::findOrFail($id);
 
         $candidature->delete();
